Adds PUT/POST/DELETE to users endpoint
