+++ rest/model/account/user.go
@@ -2,59 +2,18 @@ package account
 
 // User wraps an NS1 /account/users resource
 type User struct {
+	// Read-only fields
+	LastAccess float64 `json:"last_access"`
+
 	Name        string               `json:"name"`
 	Username    string               `json:"username"`
 	Email       string               `json:"email"`
-	Notify      NotificationSettings `json:"notify"`
 	TeamIDs     []string             `json:"teams"`
+	Notify      NotificationSettings `json:"notify"`
 	Permissions PermissionsMap       `json:"permissions"`
-	LastAccess  float64              `json:"last_access"`
 }
 
 // NotificationSettings wraps a User's "notify" attribute
 type NotificationSettings struct {
 	Billing bool `json:"billing"`
-}
-
-// PermissionsMap wraps a User's "permissions" attribute
-type PermissionsMap struct {
-	DNS        PermissionsDNS        `json:"dns"`
-	Data       PermissionsData       `json:"data"`
-	Account    PermissionsAccount    `json:"account"`
-	Monitoring PermissionsMonitoring `json:"monitoring"`
-}
-
-// PermissionsDNS wraps a User's "permissions.dns" attribute
-type PermissionsDNS struct {
-	ViewZones           bool     `json:"view_zones"`
-	ManageZones         bool     `json:"manage_zones"`
-	ZonesAllowByDefault bool     `json:"zones_allow_by_default"`
-	ZonesDeny           []string `json:"zones_deny"`
-	ZonesAllow          []string `json:"zones_allow"`
-}
-
-// PermissionsData wraps a User's "permissions.data" attribute
-type PermissionsData struct {
-	PushToDatafeeds   bool `json:"push_to_datafeeds"`
-	ManageDatasources bool `json:"manage_datasources"`
-	ManageDatafeeds   bool `json:"manage_datafeeds"`
-}
-
-// PermissionsAccount wraps a User's "permissions.account" attribute
-type PermissionsAccount struct {
-	ManageUsers           bool `json:"manage_users"`
-	ManagePaymentMethods  bool `json:"manage_payment_methods"`
-	ManagePlan            bool `json:"manage_plan"`
-	ManageTeams           bool `json:"manage_teams"`
-	ManageApikeys         bool `json:"manage_apikeys"`
-	ManageAccountSettings bool `json:"manage_account_settings"`
-	ViewActivityLog       bool `json:"view_activity_log"`
-	ViewInvoices          bool `json:"view_invoices"`
-}
-
-// PermissionsMonitoring wraps a User's "permissions.monitoring" attribute
-type PermissionsMonitoring struct {
-	ManageLists bool `json:"manage_lists"`
-	ManageJobs  bool `json:"manage_jobs"`
-	ViewJobs    bool `json:"view_jobs"`
 }
